+++ src/Tests/collisionsSuite.c
@@ -7,7 +7,7 @@
 void test_check_collision_between_ball_and_segment01(CuTest *tc){
     segment seg = {{0, 0}, {50, 0}, 0, {0, 0, 0, 0}};
     get_line_from_points(seg.A.x, seg.A.y, seg.B.x, seg.B.y, &seg.line_equation);
-    double x = 25, y = 25,
+    long double x = 25, y = 25,
            dx = 0, dy = -20,
            r = 15;
     CuAssertTrue(tc, double_abs(check_collision_between_ball_and_segment(x, y, dx, dy, r, 0, &seg) - 0.5) < eps);
@@ -16,7 +16,7 @@ void test_check_collision_between_ball_and_segment01(CuTest *tc){
 void test_check_collision_between_ball_and_segment02(CuTest *tc){
     segment seg = {{0, 0}, {50, 0}, 0, {0, 0, 0, 0}};
     get_line_from_points(seg.A.x, seg.A.y, seg.B.x, seg.B.y, &seg.line_equation);
-    double x = 25, y = 25,
+    long double x = 25, y = 25,
            dx = 60, dy = -20,
            r = 15;
     CuAssertTrue(tc, double_abs(check_collision_between_ball_and_segment(x, y, dx, dy, r, 0, &seg) - EMPTY_COLLISION_TIME) < eps);
