+++ src/stringprocessing.h
@@ -5,7 +5,7 @@
 
 char int_to_char(int);
 void int_to_str(int, char*);
-void stringify_time(char *buffer, double time);
+void stringify_time(char *buffer, long double time);
 void stringify_resolution(const ALLEGRO_DISPLAY_MODE *, char*);
 void add_char_to_name(char *buffer, char new_char, int *length, int max_length);
 void capitalize(char *a);
